Change Timer pos & Update MoneyDisplay

+++ src/logic/Shop/ShopManager.java
@@ -49,7 +49,7 @@ public class ShopManager {
             return false;
         }
 
-        player.spendMoney(item.getPrice());
+        player.spendMoney((int) item.getPrice());
         return true;
     }
